Conflicts:
	app/Http/api_routes.php
	composer.lock
	config/app.php

+++ app/Services/Logic/DeviceManager.php
@@ -2,11 +2,11 @@
 
 namespace STS\Services\Logic;
 
-use \STS\Contracts\Repository\Devices as DeviceRepository;
-use \STS\Contracts\Logic\Devices as DeviceLogic;
-
-use STS\Entities\Device;
 use STS\User;
+use Validator;
+use STS\Entities\Device;
+use STS\Contracts\Logic\Devices as DeviceLogic;
+use STS\Contracts\Repository\Devices as DeviceRepository;
 
 class DeviceManager extends BaseManager implements DeviceLogic
 {
@@ -17,32 +17,60 @@ class DeviceManager extends BaseManager implements DeviceLogic
         $this->deviceRepo = $deviceRepo;
     }
 
+    public function validator(array $data)
+    {
+        return Validator::make($data, [
+            'session_id'  => 'required|string',
+            'device_id'   => 'required|string',
+            'device_type' => 'required|string',
+            'app_version' => 'required|integer',
+        ]);
+    }
+
     public function register(User $user, array $data)
     {
-        $device = $this->deviceRepo->getDeviceBy('session_id', $data['session_id']);
-        if ($device) {
-            $this->deviceRepo->delete($device);
+        $v = $this->validator($data);
+        if ($v->fails()) {
+            $this->setErrors($v->errors());
+
+            return;
+        } else {
+            $device = $this->deviceRepo->getDeviceBy('session_id', $data['session_id']);
+            if ($device) {
+                $this->deviceRepo->delete($device);
+            }
+
+            $device = new Device();
+            $device->session_id = $data['session_id'];
+            $device->device_id = $data['device_id'];
+            $device->device_type = $data['device_type'];
+            $device->app_version = $data['app_version'];
+            $device->user_id = $user->id;
+            $this->deviceRepo->store($device);
+
+            return $device;
         }
-        
-        $device = new Device();
-        $device->session_id  = $data['session_id'];
-        $device->device_id   = $data['device_id'];
-        $device->device_type = $data['device_type'];
-        $device->app_version = $data['app_version'];
-        $device->user_id     = $user->id;
-        $this->deviceRepo->store($device);
-        return $device;
     }
 
     public function updateBySession($session_id, array $data)
     {
-        $device = $this->deviceRepo->getDeviceBy('session_id', $session_id);
-        if ($device) {
-            $device->session_id   = $data['session_id'];
-            $device->app_version   = $data['app_version'];
-            $this->deviceRepo->update($device);
+        $v = $this->validator($data);
+        if ($v->fails()) {
+            $this->setErrors($v->errors());
+
+            return;
+        } else {
+            $device = $this->deviceRepo->getDeviceBy('session_id', $session_id);
+            if ($device) {
+                $device->session_id = $data['session_id'];
+                $device->app_version = $data['app_version'];
+                $device->device_type = $data['device_type'];
+                $device->device_id = $data['device_id'];
+                $this->deviceRepo->update($device);
+            }
+
+            return $device;
         }
-        return $device;
     }
 
     public function delete($session_id)
